# OPENAPI_SWAGGER_SPEC.yaml
openapi: 3.0.0
info:
  title: Asset Management Enterprise SaaS API
  description: Complete API documentation for Asset Management system with enterprise features
  version: 1.0.0
  contact:
    name: API Support
    url: https://example.com/support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8085/api/v1
    description: Local Development
  - url: https://api.example.com/api/v1
    description: Production
paths:
...
  /assets:
    post:
      summary: Create new asset
      operationId: createAsset
      tags:
        - Assets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetRequest'
      responses:
        '201':
          description: Asset created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Insufficient permissions

    get:
      summary: List assets
      operationId: listAssets
      tags:
        - Assets
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [IN_STOCK, IN_USE, RETIRED]
        - name: departmentId
          in: query
          description: Filter by department UUID
          schema:
            type: string
        - name: categoryId
          in: query
          description: Filter by category UUID
          schema:
            type: string
        - name: page
          in: query
          description: Page number (0-indexed)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: Sort field and direction (e.g., createdAt,desc)
          schema:
            type: string
      responses:
        '200':
          description: Asset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Asset'
        '401':
          description: Unauthorized
...
  /purchase-orders:
    post:
      summary: Create purchase order
      operationId: createPurchaseOrder
      tags:
        - Purchase Orders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseOrderRequest'
      responses:
        '201':
          description: PO created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrder'
        '400':
          description: Validation error

    get:
      summary: List purchase orders
      operationId: listPurchaseOrders
      tags:
        - Purchase Orders
      security:
        - bearerAuth: []
      parameters:
        - name: departmentId
          in: query
          schema:
            type: string
        - name: supplierId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, APPROVED, REJECTED]
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: PO list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PurchaseOrder'
...
  /analytics/assets:
    get:
      summary: Get asset analytics
      operationId: getAssetAnalytics
      tags:
        - Analytics
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [month, quarter, year]
            default: month
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [status, department, category, condition]
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsData'
...
  /maintenance:
    get:
      summary: List maintenance records
      operationId: listMaintenance
      tags:
          - Maintenance
      security:
        - BearerAuth: []
      parameters:
        - name: assetId
          in: query
          required: false
          description: Filter by asset UUID
          schema:
            type: string
        - name: vendorId
          in: query
          required: false
          description: Filter by vendor UUID
          schema:
            type: string
        - name: dueBefore
          in: query
          required: false
          description: ISO-8601 date
          schema:
            type: string
      responses:
        '200':
          description: Records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceRecordDto'
        '401':
          description: Unauthorized
...
  /analytics/financial:
    get:
      summary: Financial analytics summary
      operationId: getFinancialAnalytics
      tags:
          - Analytics
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          required: false
          description: day|week|month|year (default month)
          schema:
            type: string
      responses:
        '200':
          description: Financial analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalValue:
                    type: number
                  totalDepreciation:
                    type: number
                  byMonth:
                    type: object
        '401':
          description: Unauthorized
  /analytics/purchase-orders:
    get:
      summary: Purchase order analytics
      operationId: getPurchaseOrderAnalytics
      tags:
          - Analytics
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          required: false
          description: day|week|month|year (default month)
          schema:
            type: string
      responses:
        '200':
          description: PO analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalOrders:
                    type: integer
                  totalValue:
                    type: number
                  byStatus:
                    type: object
        '401':
          description: Unauthorized
  /analytics/maintenance:
    get:
      summary: Maintenance analytics summary
      operationId: getMaintenanceAnalytics
      tags:
          - Analytics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Maintenance analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalRecords:
                    type: integer
                  completed:
                    type: integer
                  overdue:
                    type: integer
        '401':
          description: Unauthorized
  /analytics/depreciation-trends:
    get:
      summary: Monthly depreciation trend data
      operationId: getDepreciationTrends
      tags:
          - Analytics
      security:
        - BearerAuth: []
      parameters:
        - name: months
          in: query
          required: false
          description: Number of months (default 12)
          schema:
            type: integer
      responses:
        '200':
          description: Depreciation trends
          content:
            application/json:
              schema:
                type: object
                properties:
                  months:
                    type: array
                  values:
                    type: array
        '401':
          description: Unauthorized
...
  /billing/plans:
    get:
      summary: List available subscription plans
      operationId: listBillingPlans
      tags:
          - Billing
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  plans:
                    type: array
        '401':
          description: Unauthorized
  /billing/subscription:
    get:
      summary: Get current subscription
      operationId: getCurrentSubscription
      tags:
          - Billing
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Subscription
          content:
            application/json:
              schema:
                type: object
                properties:
                  planId:
                    type: string
                  status:
                    type: string
                  expiresAt:
                    type: string
                  autoRenew:
                    type: boolean
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /billing/checkout:
    post:
      summary: Initialize a checkout session
      operationId: initializeCheckout
      tags:
          - Billing
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                planId:
                  type: string
                currency:
                  type: string
                  description: USD|GHS
                successUrl:
                  type: string
                cancelUrl:
                  type: string
      responses:
        '200':
          description: Checkout
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkoutUrl:
                    type: string
                  reference:
                    type: string
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /billing/checkout/verify:
    post:
      summary: Verify a completed checkout
      operationId: verifyCheckout
      tags:
          - Billing
      security:
        - BearerAuth: []
      parameters:
        - name: reference
          in: query
          required: true
          description: Payment reference
          schema:
            type: string
      responses:
        '200':
          description: Subscription
          content:
            application/json:
              schema:
                type: object
                properties:
                  planId:
                    type: string
                  status:
                    type: string
                  expiresAt:
                    type: string
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /billing/subscription/auto-renew:
    patch:
      summary: Toggle auto-renew setting
      operationId: setAutoRenew
      tags:
          - Billing
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                autoRenew:
                  type: boolean
      responses:
        '200':
          description: Updated subscription
          content:
            application/json:
              schema:
                type: object
                properties:
                  autoRenew:
                    type: boolean
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AssetRequest:
      type: object
      required:
        - name
        - assetTag
        - categoryId
        - departmentId
        - purchaseCost
        - currency
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        assetTag:
          type: string
        assetType:
          type: string
        serialNumber:
          type: string
        categoryId:
          type: string
          format: uuid
        departmentId:
          type: string
          format: uuid
        locationId:
          type: string
          format: uuid
        supplierId:
          type: string
          format: uuid
        purchaseCost:
          type: number
          format: double
        currency:
          type: string
          enum: [USD, EUR, GBP]
        status:
          type: string
          enum: [IN_STOCK, IN_USE, RETIRED]
        condition:
          type: string
          enum: [POOR, FAIR, GOOD, EXCELLENT]
    Asset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        assetTag:
          type: string
        status:
          type: string
        condition:
          type: string
        purchaseCost:
          type: number
        currentBookValue:
          type: number
        organisationId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PurchaseOrderRequest:
      type: object
      required:
        - poNumber
        - totalAmount
        - currency
        - departmentId
        - supplierId
      properties:
        poNumber:
          type: string
        totalAmount:
          type: number
          format: double
        currency:
          type: string
        departmentId:
          type: string
          format: uuid
        supplierId:
          type: string
          format: uuid
        remarks:
          type: string
        status:
          type: string
          enum: [DRAFT, APPROVED, REJECTED]
    PurchaseOrder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        poNumber:
          type: string
        totalAmount:
          type: number
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AnalyticsData:
      type: object
      properties:
        period:
          type: string
        groupBy:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              count:
                type: integer
              value:
                type: number
              percentage:
                type: number
...
    ValidationError:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string